mark parallel test as big
