+++ src/dbnode/storage/series/series_test.go
@@ -24,7 +24,6 @@ import (
 	"errors"
 	"fmt"
 	"io"
-	"sync"
 	"testing"
 	"time"
 
@@ -848,55 +847,6 @@ func TestSeriesWriteReadFromTheSameBucket(t *testing.T) {
 	require.NoError(t, err)
 
 	require.Equal(t, 3, len(values))
-}
-
-// TestSeriesWriteReadParallel is a regression test that was added to capture panics that might
-// arise when many parallel writes and reads are happening at the same time.
-func TestSeriesWriteReadParallel(t *testing.T) {
-	var (
-		numWorkers        = 100
-		numStepsPerWorker = numWorkers * 100
-		opts              = newSeriesTestOptions()
-		curr              = time.Now()
-		series            = NewDatabaseSeries(ident.StringID("foo"), ident.Tags{}, opts).(*dbSeries)
-	)
-
-	_, err := series.Bootstrap(nil)
-	assert.NoError(t, err)
-
-	ctx := context.NewContext()
-	defer ctx.Close()
-
-	wg := sync.WaitGroup{}
-	wg.Add(1)
-	go func() {
-		for i := 0; i < numStepsPerWorker; i++ {
-			wasWritten, err := series.Write(
-				ctx, curr.Add(time.Duration(i)*time.Nanosecond), float64(i), xtime.Second, nil)
-			if err != nil {
-				panic(err)
-			}
-			if !wasWritten {
-				panic("write failed")
-			}
-		}
-		wg.Done()
-	}()
-
-	// Outer loop so that reads are competing with other reads, not just writes.
-	for j := 0; j < numWorkers; j++ {
-		wg.Add(1)
-		go func() {
-			for i := 0; i < numStepsPerWorker; i++ {
-				_, err := series.ReadEncoded(ctx, curr.Add(-5*time.Minute), curr.Add(time.Minute))
-				if err != nil {
-					panic(err)
-				}
-			}
-			wg.Done()
-		}()
-	}
-	wg.Wait()
 }
 
 func TestSeriesCloseNonCacheLRUPolicy(t *testing.T) {
